feat(hydra-indexer): add migration for substrate_block
affects: @dzlzv/hydra-common, @dzlzv/hydra-indexer
