+++ packages/hydra-indexer/src/db/ormconfig.ts
@@ -1,25 +1,6 @@
-import { ConnectionOptions } from 'typeorm'
-import { SnakeNamingStrategy } from '@dzlzv/hydra-db-utils'
-import { SubstrateEventEntity, SubstrateExtrinsicEntity } from '../entities'
-import { getDBConfig } from '../node'
+import config from './dbconfig'
 
-const config: () => ConnectionOptions = () => {
-  const conf = getDBConfig()
-  return {
-    type: 'postgres',
-    host: conf.DB_HOST,
-    port: conf.DB_PORT,
-    username: conf.DB_USER,
-    password: conf.DB_PASS,
-    database: conf.DB_NAME,
-    entities: [SubstrateEventEntity, SubstrateExtrinsicEntity],
-    migrations: ['./**/migrations/*.js'],
-    cli: {
-      migrationsDir: 'migrations',
-    },
-    logging: conf.DB_LOGGING,
-    namingStrategy: new SnakeNamingStrategy(),
-  } as ConnectionOptions
-}
+const ormconfig = config()
 
-export default config
+// export dbconfig as default so that it can be use with typeorm-cli
+export default ormconfig
